requirements.txt
einops==0.8.0
faiss==1.7.3
flash_attn==2.5.8
numpy==1.21.4
PyYAML==6.0.1
PyYAML==6.0.1
tiktoken==0.7.0
torch==1.13.1
tqdm==4.62.3
transformers==4.34.0
transformers_stream_generator==0.0.4
